adds new github commands page for best practice / most use command
The intend of this page is providing a list of best practices and most
uses of commands, methods or solutions when handling with git or GitHub.

The first issue describes the case of updating the forked repository
with the upstream's data. Additionally, it will show some hints in
other tools when available.